+++ epicore_utils/modules/parse_input.py
@@ -211,7 +211,7 @@ def group_repetitive(start: list[int], end: list[int], pep: list[str], acc, idx,
         end positions that are part of repetitive regions the lowest start 
         position and highest end position is kept for each repetitive region. 
     """
-     # TODO have to be sorted
+    #lodo have to be sorted
 
 
     current = -1
@@ -328,12 +328,11 @@ def prot_pep_link(peptides_df: pd.DataFrame, seq_column: str, protacc_column: st
         proteins_df = proteins_df.with_columns(pl.col('peptide_index').cast(pl.List(pl.List(pl.Int64))))
         proteins_df = proteins_df.with_columns(pl.col('sample').cast(pl.List(pl.String)))
         proteins_df = proteins_df.with_columns(pl.col('condition').cast(pl.List(pl.String)))
-        
-        block_size = max(len(proteins_df) // 10,1)
-        with get_context('spawn').Pool(10) as pool:
-            chunk_dfs = pool.starmap(group_repetitive_chunk,[(proteins_df,chunk*block_size,(chunk+1)*block_size if chunk < 9 else len(proteins_df)) for chunk in range(10)])
+
+        block_size = max(len(proteins_df) // 20,1)
+        with get_context('spawn').Pool(20) as pool:
+            chunk_dfs = pool.starmap(group_repetitive_chunk,[(proteins_df,chunk*block_size,(chunk+1)*block_size if chunk < 19 else len(proteins_df)) for chunk in range(20)])
         proteins_df = pl.concat(chunk_dfs)
-        
         proteins_df = proteins_df.with_columns(pl.col('repetitive').list.get(0).alias('start'))
         proteins_df = proteins_df.with_columns(pl.col('repetitive').list.get(1).alias('end'))
         proteins_df = proteins_df.with_columns(pl.col('repetitive').list.get(2).alias('peptide_index'))